refactor(core): extract types
